+++ tests/testthat/test-export_as_docx.R
@@ -15,7 +15,7 @@ test_that("export_as_docx works thanks to tt_to_flextable", {
   # Get the flextable
   flex_tbl <- tt_to_flextable(tbl, titles_as_header = TRUE, integrate_footers = TRUE)
 
-  doc_file <- tempfile(fileext = ".docx")
+  doc_file <- tempfile(tmpdir = tempdir(check = TRUE), fileext = ".docx")
 
   expect_silent(export_as_docx(tbl,
     file = doc_file, doc_metadata = list("title" = "meh"),
@@ -43,7 +43,7 @@ test_that("export_as_docx produces a warning if manual column widths are used", 
     analyze("Petal.Length")
   tbl <- build_table(lyt, iris)
 
-  doc_file <- tempfile(fileext = ".docx")
+  doc_file <- tempfile(tmpdir = tempdir(check = TRUE), fileext = ".docx")
 
   # Get the flextable
   expect_warning(
@@ -65,8 +65,54 @@ test_that("export_as_docx works thanks to tt_to_flextable", {
     add_trailing_sep = "ARM"
   )
 
-  doc_file <- tempfile(fileext = ".docx")
+  doc_file <- tempfile(tmpdir = tempdir(check = TRUE), fileext = ".docx")
   expect_no_error(
     out <- export_as_docx(lsting, doc_file, titles_as_header = TRUE, integrate_footers = TRUE)
+  )
+})
+
+
+test_that("Getting correct template file", {
+  root <- system.file(package = "rtables.officer")
+  expect_equal(
+    .get_template_file(section_properties_default(page_size = "A4", orientation = "portrait")),
+    file.path(root, "docx_templates/a4_portrait.docx")
+  )
+
+  expect_equal(
+    .get_template_file(section_properties_default(page_size = "A4", orientation = "landscape")),
+    file.path(root, "docx_templates/a4_landscape.docx")
+  )
+
+  expect_equal(
+    .get_template_file(section_properties_default(page_size = "letter", orientation = "portrait")),
+    file.path(root, "docx_templates/letter_portrait.docx")
+  )
+
+  expect_equal(
+    .get_template_file(section_properties_default(page_size = "letter", orientation = "landscape")),
+    file.path(root, "docx_templates/letter_landscape.docx")
+  )
+
+  # Warnings
+  spd <- section_properties_default(page_size = "A4", orientation = "portrait")
+  spd$page_size$width <- 6
+  expect_warning(
+    void <- .get_template_file(spd),
+    "Adding page numbers is supported only A4 and letter size."
+  )
+  spd <- section_properties_default(page_size = "A4", orientation = "landscape")
+  spd$page_size$width <- 6
+  expect_warning(
+    void <- .get_template_file(spd),
+    "Adding page numbers is supported only A4 and letter size."
+  )
+
+  # Error
+  spd <- section_properties_default(page_size = "A4", orientation = "landscape")
+  spd$page_size$orient <- "orient"
+  expect_error(
+    void <- .get_template_file(spd),
+    "Adding page numbers is supported only for landscape and portrait orientation."
   )
 })